+++ CHANGELOG.md
@@ -186,3 +186,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Mailing Black List
 - Sort Mailing Black List by first and last name
 - Results
+- DRN in exporter label